feat: get median wasm with test
