+++ get-median-wasm/test.json
@@ -1,0 +1,20 @@
+{
+    "Column_1": [
+        8.1,
+        6.1,
+        3,
+        3,
+        7,
+        1,
+        9
+    ],
+    "Column_2": [
+        8.1,
+        6.1,
+        5,
+        3,
+        7,
+        7,
+        9
+    ]
+}
